requirements.txt
Django==2.0.2
mysqlclient
django-cors-headers
django-xadmin